+++ Runtime/ViewJson.cs
@@ -1,0 +1,9 @@
+﻿namespace At.Ac.FhStp.ViewJson
+{
+    /// <summary>
+    /// This class contains functions for converting json to Unity UGUI
+    /// </summary>
+    public static class ViewJson
+    {
+    }
+}